Move js files to service
